#2701:  * Reusing similiarty calculator in BinaryMemPR  * Fixing distance calculation for linear linkage and LinearLinkageMemPR  * Small changes to base algorithm  * Added biased model trainer for permutation (rank and fitness)  * Fixing best known quality calculation for GCP
git-svn-id: https://src.heuristiclab.com/svn/core/branches/MemPRAlgorithm@14496 2abd9481-f8db-48e9-bd25-06bc13291c1b

+++ HeuristicLab.Problems.Instances/3.3/Types/GCPData.cs
@@ -47,8 +47,9 @@ namespace HeuristicLab.Problems.Instances {
     /// </summary>
     public int[] BestKnownColoring { get; set; }
     /// <summary>
-    /// Optional! The quality value of the <see cref="BestKnownColoring"/>
+    /// Optional! The least amount of colors that would not result in conflicts.
+    /// The amount of colors in <see cref="BestKnownColoring"/> if it is given as well.
     /// </summary>
-    public double? BestKnownColors { get; set; }
+    public int? BestKnownColors { get; set; }
   }
 }
